+++ CHANGELOG.md
@@ -51,6 +51,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+- [#480](https://github.com/alleslabs/celatone-frontend/pull/480) Update chain registry package and add attach funds in code snippet
 - [#474](https://github.com/alleslabs/celatone-frontend/pull/474) Refactor stored and saved codes pages
 - [#468](https://github.com/alleslabs/celatone-frontend/pull/468) Add breadcrumb to pool id page
 - [#466](https://github.com/alleslabs/celatone-frontend/pull/466) Fix developer mode alert
@@ -70,7 +71,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Bug fixes
 
-- [#480](https://github.com/alleslabs/celatone-frontend/pull/480) Fix attach funds in code snippet
 - [#476](https://github.com/alleslabs/celatone-frontend/pull/476) Fix terra rpc
 - [#473](https://github.com/alleslabs/celatone-frontend/pull/473) Fix wording on upload code modal
 - [#465](https://github.com/alleslabs/celatone-frontend/pull/465) Fix wrong access queryKey in validator info